Change repo structure, refactor CADA-VAE, add Trainer abstraction
