+++ datasets/README.md
@@ -1,0 +1,6 @@
+Using built-in datasets
+=======================
+
+ZeroShotEval is provided with 2 build-in datasest:
+ - AWA2
+ - CUB
