+++ zeroshoteval/feature_extraction/attributes_embedding_extractor.py
