+++ zeroshoteval/feature_extraction/image_embedding_extractor.py
