+++ zeroshoteval/feature_extraction/modalities_feature_extractor.py
